requirements.txt
# Web UI
streamlit==1.37.1
pyyaml==6.0.2

# LangChain + LLM
langchain==0.2.11
langchain-community==0.2.10

# Vector search
faiss-cpu==1.9.0      # compatible with Python 3.12 + Windows
sentence-transformers==3.0.1

# NLP / LLMs
transformers==4.43.3
torch==2.3.1



# Utils
pandas==2.2.2
